refactor: pageに引数を渡すのは定義と処理が遠いのでindexのほうに移した
引数の型は現状使わないが今後コンポーネントわけしていくのでコメントアウトしておく
