+++ resources/js/pages/top/index.tsx
@@ -1,47 +1,9 @@
-import type { ColumnDef } from '@tanstack/react-table';
 import { Page } from './page';
 import { createRoot } from 'react-dom/client';
 
 const domNode = document.getElementById('top');
 
-export type Payment = {
-  id: string;
-  amount: number;
-  status: 'pending' | 'processing' | 'success' | 'failded';
-  email: string;
-};
-
-export const payments: Payment[] = [
-  {
-    id: '728ed52f',
-    amount: 100,
-    status: 'pending',
-    email: 'm@example.com',
-  },
-  {
-    id: '489e1d42',
-    amount: 125,
-    status: 'processing',
-    email: 'example@gmail.com',
-  },
-];
-
-const columns: ColumnDef<Payment>[] = [
-  {
-    accessorKey: 'status',
-    header: 'Status',
-  },
-  {
-    accessorKey: 'email',
-    header: 'Email',
-  },
-  {
-    accessorKey: 'amount',
-    header: 'Amount',
-  },
-];
-
 if (domNode) {
   const root = createRoot(domNode);
-  root.render(<Page columns={columns} data={payments} />);
+  root.render(<Page />);
 }
